requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
fastapi==0.75.0
h11==0.13.0
idna==3.3
pydantic==1.9.0
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
